requirements.txt
streamlit
pandas
plotly
reportlab
python-dotenv
langchain-groq
langchain-core
pydantic
requests